Split image gallery selection hook
