+++ tests/image-workspace-split-contract.test.ts
@@ -9,6 +9,7 @@ const workspacePath = path.join(imageDir, 'ImageWorkspace.tsx');
 const composerPersistenceHookPath = path.join(imageDir, '_hooks/useImageComposerPersistence.ts');
 const queryHydrationHookPath = path.join(imageDir, '_hooks/useImageWorkspaceQueryHydration.ts');
 const generationRunnerHookPath = path.join(imageDir, '_hooks/useImageGenerationRunner.ts');
+const gallerySelectionHookPath = path.join(imageDir, '_hooks/useImageGallerySelection.ts');
 
 const splitFiles = [
   '_components/ImageAuthGateModal.tsx',
@@ -19,6 +20,7 @@ const splitFiles = [
   '_hooks/useImageWorkspaceDesktopLayout.ts',
   '_hooks/useImageWorkspaceQueryHydration.ts',
   '_hooks/useImageGenerationRunner.ts',
+  '_hooks/useImageGallerySelection.ts',
   '_lib/image-workspace-character-references.ts',
   '_lib/image-workspace-copy.ts',
   '_lib/image-workspace-history.ts',
@@ -36,6 +38,7 @@ test('image workspace foundations are split from the route orchestrator', () => 
   const composerPersistenceHookSource = readFileSync(composerPersistenceHookPath, 'utf8');
   const queryHydrationHookSource = readFileSync(queryHydrationHookPath, 'utf8');
   const generationRunnerHookSource = readFileSync(generationRunnerHookPath, 'utf8');
+  const gallerySelectionHookSource = readFileSync(gallerySelectionHookPath, 'utf8');
 
   for (const file of splitFiles) {
     statSync(path.join(imageDir, file));
@@ -49,8 +52,8 @@ test('image workspace foundations are split from the route orchestrator', () => 
   assert.match(source, /from '\.\/_hooks\/useImageWorkspaceDesktopLayout'/);
   assert.match(source, /from '\.\/_hooks\/useImageWorkspaceQueryHydration'/);
   assert.match(source, /from '\.\/_hooks\/useImageGenerationRunner'/);
+  assert.match(source, /from '\.\/_hooks\/useImageGallerySelection'/);
   assert.match(source, /from '\.\/_lib\/image-workspace-copy'/);
-  assert.match(source, /from '\.\/_lib\/image-workspace-utils'/);
 
   assert.doesNotMatch(source, /const DEFAULT_COPY: ImageWorkspaceCopy =/);
   assert.doesNotMatch(source, /function ImageLibraryModal\(/);
@@ -71,6 +74,10 @@ test('image workspace foundations are split from the route orchestrator', () => 
   assert.doesNotMatch(source, /runImageGeneration/, 'generation submission belongs in useImageGenerationRunner');
   assert.doesNotMatch(source, /readBrowserSession/, 'auth preflight belongs in useImageGenerationRunner');
   assert.doesNotMatch(source, /validateGptImage2CustomImageSize/, 'custom size validation belongs in useImageGenerationRunner');
+  assert.doesNotMatch(source, /const fetchSnapshot =/, 'gallery snapshot hydration belongs in useImageGallerySelection');
+  assert.doesNotMatch(source, /const previewUrls =/, 'fallback gallery image derivation belongs in useImageGallerySelection');
+  assert.doesNotMatch(source, /findImageEngine/, 'gallery engine default selection belongs in useImageGallerySelection');
+  assert.doesNotMatch(source, /authFetch/, 'gallery API snapshot lookup belongs in useImageGallerySelection');
 
   assert.match(composerPersistenceHookSource, /export function useImageComposerPersistence/);
   assert.match(composerPersistenceHookSource, /parsePersistedImageComposerState/);
@@ -87,7 +94,12 @@ test('image workspace foundations are split from the route orchestrator', () => 
   assert.match(generationRunnerHookSource, /from '\.\.\/_lib\/image-workspace-history'/);
   assert.match(generationRunnerHookSource, /buildPendingGroup/);
   assert.match(generationRunnerHookSource, /buildCompletedGroup/);
+  assert.match(gallerySelectionHookSource, /export function useImageGallerySelection/);
+  assert.match(gallerySelectionHookSource, /findImageEngine/);
+  assert.match(gallerySelectionHookSource, /getAspectRatioOptions/);
+  assert.match(gallerySelectionHookSource, /authFetch/);
+  assert.match(gallerySelectionHookSource, /const previewUrls =/);
 
   const lineCount = source.split('\n').length;
-  assert.ok(lineCount <= 950, `ImageWorkspace should stay below 950 lines after generation runner extraction, got ${lineCount}`);
+  assert.ok(lineCount <= 830, `ImageWorkspace should stay below 830 lines after gallery selection extraction, got ${lineCount}`);
 });
